requirements.txt
requests
pdfminer3k
googletrans